PatrolRoute Pacing
Added a movingForward boolean to keep track of the patrol direction.

When the shark reaches the last waypoint, set movingForward to false, and the shark starts moving to the previous waypoint.
When the shark reaches the first waypoint,  set movingForward to true again, and it resumes patrolling forward.
We also added a rotation step that orients the shark to face its next target waypoint for a more realistic movement.